Add macro to execute the current cell in MATLAB
Needs more work, currently it will actually select the entire cell instead of leaving the cursor at its current location; if you run the macro twice in a row it will move to the next cell instead of executing the current cell twice. Probably needs to be written as an actual command.

The execute command's script has been moved into a stand alone script so it can be used in this command as well.